ai_detection_service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=2.0.0
python-multipart>=0.0.6
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
